+++ src/wallcrossingaction.cc
@@ -13,7 +13,7 @@ namespace Smash {
 
 ActionList WallCrossActionsFinder::find_actions_in_cell(
                    const ParticleList &plist,
-                   float t_max) const {
+                   double t_max) const {
   std::vector<ActionPtr> actions;
   for (const ParticleData &p : plist) {
     const ThreeVector& r = p.position().threevec();
